+++ db/migrations/20170717112700_create_blobs.rb
@@ -1,0 +1,18 @@
+Hanami::Model.migration do
+  up do
+    execute 'CREATE EXTENSION IF NOT EXISTS "uuid-ossp"'
+    create_table :blobs do
+      primary_key :id, 'uuid', null: false, default: Hanami::Model::Sql.function(:uuid_generate_v4)
+
+      column :format, String, null: false
+
+      column :created_at, DateTime, null: false
+      column :updated_at, DateTime, null: false
+    end
+  end
+
+  down do
+    drop_table :blobs
+    execute 'DROP EXTENSION IF EXISTS "uuid-ossp"'
+  end
+end