+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+* Use UUID as primary key for Blobs (breaking change, please drop db before)
+* Add a Utils helper to check UUID format
 * Add Blob entities to represent data and metadata
 * Add a search endpoint on data
 * Add project to Travis CI